mdec_2025/requirements.txt
accelerate
diffusers>=0.31.0
numpy
Pillow
scipy
torch
torchvision
transformers
tqdm
